backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
channels==4.2.2
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.6.1
PyJWT==2.9.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uvicorn==0.34.0
whitenoise==6.9.0
